// web/src/api/generated/model/upload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */

export interface Upload {
  declared_size: number;
  expires_at: string;
  failure_code?: string;
  failure_message?: string;
  id: string;
  node_id?: string;
  offset: number;
  state: string;
  upload_url: string;
  version_id?: string;
}